requirements.txt
# Skate Physics Preserver - Colab T4 (Python 3.10)
# Pin versions for reproducibility

# PyTorch (Colab has CUDA 11.8 or 12.x - use default)
torch>=2.0.0
torchvision>=0.15.0

# SAM 2.1
sam2 @ git+https://github.com/facebookresearch/sam2.git@main

# DWPose / Pose
rtmlib>=0.0.15

# ONNX Runtime (GPU for Colab)
onnxruntime-gpu>=1.16.0

# Video / Image
opencv-python>=4.8.0
numpy<2.0,>=1.24.0

# ComfyUI API client
requests>=2.28.0
websocket-client>=1.5.0

# Utilities
tqdm>=4.65.0
Pillow>=9.0.0
hydra-core>=1.3.0
huggingface_hub>=0.19.0
